[refactor] `ProjectExtensions`
