+++ build-logic/convention/src/main/kotlin/ir/composenews/ProjectExtensions.kt
@@ -1,25 +1,28 @@
-/*
- * Copyright 2023 The Android Open Source Project
- *
- *   Licensed under the Apache License, Version 2.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.apache.org/licenses/LICENSE-2.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- */
-
 package ir.composenews
 
+import com.android.build.api.dsl.ApplicationExtension
+import com.android.build.gradle.LibraryExtension
+import io.gitlab.arturbosch.detekt.extensions.DetektExtension
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
 import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.getByType
 
 val Project.libs
     get(): VersionCatalog = extensions.getByType<VersionCatalogsExtension>().named("libs")
+
+inline fun Project.androidGradle(crossinline configure: LibraryExtension.() -> Unit) =
+    extensions.configure<LibraryExtension> {
+        configure()
+    }
+
+inline fun Project.detektGradle(crossinline configure: DetektExtension.() -> Unit) =
+    extensions.configure<DetektExtension> {
+        configure()
+    }
+
+inline fun Project.applicationGradle(crossinline configure: ApplicationExtension.() -> Unit) =
+    extensions.configure<ApplicationExtension> {
+        configure()
+    }
